requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
llama-index>=0.12.0
llama-index-llms-gemini
python-dotenv>=1.0.0
openai>=1.0.0
google-generativeai>=0.5.0
pytesseract>=0.3.10
pdf2image>=1.16.3
pdfplumber>=0.10.3